+++ Tests.Analytics/UnitTesting.cs
@@ -40,7 +40,7 @@ namespace Tests.Analytics
 			{
 				var data = new[] { 0M, 0.1M, 0.6M, 0.8M, 1.1M, 2.2M, 2.4M, };
 				var distribution = Stats.CalculateDistribution(data, 0.5M);
-				var sample = new SortedDictionary<decimal, decimal>
+				var sample = new SortedDictionary<decimal, int>
 				    {
 						{ 0, 1 },
 				        { 0.5M, 1 },
@@ -54,7 +54,7 @@ namespace Tests.Analytics
 			{
 				var data = new[] { 0M, 0.1M, 0.2M, 0.21M, 0.25M };
 				var distribution = Stats.CalculateDistribution(data, 0.2M);
-				var sample = new SortedDictionary<decimal, decimal>
+				var sample = new SortedDictionary<decimal, int>
 				    {
 				        { 0, 1 },
 						{ 0.2M, 2 },
